* Changed SCF_NO_CLEAR to SCF_PERMANENT, to makes a sense for 'human' friendly flag name. * Updated sc_config.txt documentation (maybe just as temporary explanation?). * Added flags for bonus_script's flags, imitating sc_config flags. * misc...

+++ src/map/pc.h
@@ -587,7 +587,7 @@ struct map_session_data {
 		struct script_code *script;
 		char script_str[MAX_BONUS_SCRIPT_LENGTH]; //Used for comparing and storing on table
 		uint32 tick;
-		uint8 flag;
+		uint16 flag;
 		char type; //0 - Ignore; 1 - Buff; 2 - Debuff
 		int16 icon;
 		int tid;
